Make llama stack build not create a new conda by default
