+++ llama_stack/cli/stack/run.py
@@ -38,6 +38,11 @@ class StackRun(Subcommand):
             default=int(os.getenv("LLAMA_STACK_PORT", 5000)),
         )
         self.parser.add_argument(
+            "--image-name",
+            type=str,
+            help="Name of the image to run. Defaults to the current conda environment",
+        )
+        self.parser.add_argument(
             "--disable-ipv6",
             action="store_true",
             help="Disable IPv6 support",
@@ -53,8 +58,11 @@ class StackRun(Subcommand):
 
     def _run_stack_run_cmd(self, args: argparse.Namespace) -> None:
         import importlib.resources
+        import json
+        import subprocess
 
         import yaml
+        from termcolor import cprint
 
         from llama_stack.distribution.build import ImageType
         from llama_stack.distribution.configure import parse_and_maybe_upgrade_config
@@ -99,11 +107,11 @@ class StackRun(Subcommand):
 
         if not config_file.exists():
             self.parser.error(
-                f"File {str(config_file)} does not exist. Please run `llama stack build` to generate (and optionally edit) a run.yaml file"
+                f"File {str(config_file)} does not exist.\n\nPlease run `llama stack build` to generate (and optionally edit) a run.yaml file"
             )
             return
 
-        print(f"Using config file: {config_file}")
+        print(f"Using run configuration: {config_file}")
         config_dict = yaml.safe_load(config_file.read_text())
         config = parse_and_maybe_upgrade_config(config_dict)
 
@@ -114,13 +122,52 @@ class StackRun(Subcommand):
             )
             run_args = [script, config.docker_image]
         else:
+            current_conda_env = os.environ.get("CONDA_DEFAULT_ENV")
+            image_name = args.image_name or current_conda_env
+            if not image_name:
+                cprint(
+                    "No current conda environment detected, please specify a conda environment name with --image-name",
+                    color="red",
+                )
+                return
+
+            def get_conda_prefix(env_name):
+                # Get conda environments info
+                conda_env_info = json.loads(
+                    subprocess.check_output(
+                        ["conda", "info", "--envs", "--json"]
+                    ).decode()
+                )
+                envs = conda_env_info["envs"]
+                for envpath in envs:
+                    if envpath.endswith(env_name):
+                        return envpath
+                return None
+
+            print(f"Using conda environment: {image_name}")
+            conda_prefix = get_conda_prefix(image_name)
+            if not conda_prefix:
+                cprint(
+                    f"Conda environment {image_name} does not exist.",
+                    color="red",
+                )
+                return
+
+            build_file = Path(conda_prefix) / "llamastack-build.yaml"
+            if not build_file.exists():
+                cprint(
+                    f"Build file {build_file} does not exist.\n\nPlease run `llama stack build` or specify the correct conda environment name with --image-name",
+                    color="red",
+                )
+                return
+
             script = (
                 importlib.resources.files("llama_stack")
                 / "distribution/start_conda_env.sh"
             )
             run_args = [
                 script,
-                config.conda_env,
+                image_name,
             ]
 
         run_args.extend([str(config_file), str(args.port)])
@@ -129,13 +176,17 @@ class StackRun(Subcommand):
 
         for env_var in args.env:
             if "=" not in env_var:
-                self.parser.error(
-                    f"Environment variable '{env_var}' must be in KEY=VALUE format"
+                cprint(
+                    f"Environment variable '{env_var}' must be in KEY=VALUE format",
+                    color="red",
                 )
                 return
             key, value = env_var.split("=", 1)  # split on first = only
             if not key:
-                self.parser.error(f"Environment variable '{env_var}' has empty key")
+                cprint(
+                    f"Environment variable '{env_var}' has empty key",
+                    color="red",
+                )
                 return
             run_args.extend(["--env", f"{key}={value}"])
