--- llama_stack/providers/inline/inference/meta_reference/quantization/scripts/build_conda.sh
@@ -1,36 +1,0 @@
-#!/bin/bash
-
-# Copyright (c) Meta Platforms, Inc. and affiliates.
-# All rights reserved.
-#
-# This source code is licensed under the terms described in the LICENSE file in
-# the root directory of this source tree.
-
-if [[ $# -ne 1 ]]; then
-    echo "Error: Please provide the name of CONDA environment you wish to create"
-    exit 1
-fi
-
-ENV_NAME=$1
-
-set -eu
-eval "$(conda shell.bash hook)"
-
-echo "Will build env (or overwrite) named '$ENV_NAME'"
-
-set -x
-
-run_build() {
-    # Set up the conda environment
-    yes | conda remove --name $ENV_NAME --all
-    yes | conda create -n $ENV_NAME python=3.10
-    conda activate $ENV_NAME
-
-    # PT nightly
-    pip install --pre torch --index-url https://download.pytorch.org/whl/nightly/cu121
-
-    # install dependencies for `llama-agentic-system`
-    pip install -r fp8_requirements.txt
-}
-
-run_build
